Added certificate support for Android
 - Install-tool checks for certificate node
 - Alias, password and alias password are optional
 - Install-tool now defines "debug" when debug enabled

Here is a full example:

<certificate path="C:\Users\Joshua\mystore.keystore" alias="mystore" password="mypassword" alias_password="mypassword" />


Here is a simple example:

<certificate path="C:\Users\Joshua\mystore.keystore" unless="debug" />


If passwords are not defined, it prompts when you compile. If you define password, alias password is automatically set to the same. If you do not define alias, it is automatically set to the filename (minus extension) of your keystore file.


An example command-line for running keytool to generate your own keystore is available here:

http://developer.android.com/guide/publishing/app-signing.html#cert

git-svn-id: http://nekonme.googlecode.com/svn/trunk@912 1509560c-5e2a-0410-865c-31c25e1cfdef