Added automatic configuration for Android
Usage:

nme setup android


Results:

It will prompt for the path to the Android SDK, Android NDK, Apache Ant and Java JDK. If you hit enter, it will use the current value. It can also detect if the current value is a system variable, in which case it will not write the value to your .hxcpp_config.xml file.

git-svn-id: http://nekonme.googlecode.com/svn/trunk@1042 1509560c-5e2a-0410-865c-31c25e1cfdef