Initial support for improved extensions
Originally, the <extension> tag assumed that there was a single NDLL with the same name as the extension, as well as classes in the root path of the extension folder.

Now the build process should look for an XML file with the same name as the extension ("test" would search for "test.xml") and will use it as an include. This means that extensions should (theoretically) be able to define classpaths, NDLLs and other resources just as flexibly as a regular NMML project file.

This opens the door to multiple NDLLs per extension, different classpaths, unique defines, or it could include additional assets. I haven't added support for treating a haxelib as an extension, but that's a natural evolution of this feature

git-svn-id: http://nekonme.googlecode.com/svn/trunk@1113 1509560c-5e2a-0410-865c-31c25e1cfdef